requirements.txt
flask
streamlit
requests

# Vision
ultralytics
opencv-python
pillow
numpy
easyocr

# Agentic AI
langchain
langgraph
langchain-core
langchain-community
langchain-google-genai
google-generativeai

# Persistence & utils
sqlite-utils
python-dotenv
pandas
loguru